// src/proto/DmSegMobileReply.ts
export enum DmColorfulType {
  /** NoneType - 无 */
  NoneType = 0,
  /** VipGradualColor - 大会员渐变色 */
  VipGradualColor = 60001,
  UNRECOGNIZED = -1,
}

/** 弹幕来源 */
export enum DmFromType {
  /** DmFromUnknown - 未知 */
  DmFromUnknown = 0,
  /** DmFromNormal - 一般 */
  DmFromNormal = 1,
  /** DmFromCmd - 指令 */
  DmFromCmd = 2,
  /** DmFromLive - 直播 */
  DmFromLive = 3,
  UNRECOGNIZED = -1,
}

/** 分段弹幕返回 */
export interface DmSegMobileReply {
  /** elems 弹幕列表 */
  elems: DanmakuElem[];
  /**
   * state 弹幕区状态
   * | 0 | 1 |
   * | :=: | :=: |
   * | 打开 | 关闭弹幕区 |
   */
  state?:
    | number
    | undefined;
  /** ai_flag ai打分 */
  aiFlag?:
    | DanmakuAIFlag
    | undefined;
  /** 彩色弹幕 */
  colorfulSrc: DmColorful[];
}

/** 弹幕实例 */
export interface DanmakuElem {
  /** 弹幕id */
  id: bigint;
  /** 弹幕位置：/ms */
  progress?:
    | number
    | undefined;
  /**
   * 弹幕类型
   * | 1 | 4 | 5 | 6 | 7 | 8 | 9 |
   * | :-: | :-: | :-: | :-: | :-: | :-: | :-: |
   * | 普通 | 底部 | 顶部 | 逆向 | 高级 | 代码 | BAS |
   */
  mode: number;
  /** 弹幕字体 */
  fontsize: number;
  /** 弹幕颜色 */
  color: number;
  /** 弹幕发送者crc32哈希 */
  midHash: string;
  /** 弹幕文本内容 */
  content?:
    | string
    | undefined;
  /** 弹幕发送时间  时间戳 */
  ctime: bigint;
  /** 弹幕权重 越高显示优先级越高 */
  weight?:
    | number
    | undefined;
  /** 弹幕动作 */
  action?:
    | string
    | undefined;
  /**
   * 弹幕池
   * | 0 | 1 | 2 |
   * | :-: | :-: | :-: |
   * | 普通弹幕 | 字幕弹幕 | 特殊弹幕 |
   */
  pool?:
    | number
    | undefined;
  /** 弹幕id_str */
  idStr?:
    | string
    | undefined;
  /**
   * 弹幕属性位
   * | 0 | 1 | 2 |
   * | :=: | :=: | :=: |
   * | 保护弹幕 | 直播弹幕 | 高赞弹幕 |
   */
  attr?:
    | number
    | undefined;
  /** 点赞数 */
  likeCount?: bigint | undefined;
  animation?:
    | string
    | undefined;
  /** 弹幕渐变色 */
  colorful?:
    | DmColorfulType
    | undefined;
  /** 弹幕oid */
  oid?:
    | bigint
    | undefined;
  /** 弹幕来源 */
  dmFrom?: DmFromType | undefined;
}

/** AI云屏蔽弹幕的配置 */
export interface DanmakuAIFlag {
  /** 云屏蔽等级 */
  dmFlags: DanmakuFlag[];
}

/** 云屏蔽等级 */
export interface DanmakuFlag {
  /** dmid 弹幕id */
  dmid: bigint;
  /** 弹幕分 */
  flag: number;
}

/** 彩色弹幕 */
export interface DmColorful {
  /** 类型 */
  type: DmColorfulType;
  /** 图片 */
  src: string;
}

function createBaseDmSegMobileReply(): DmSegMobileReply {
  return { elems: [], colorfulSrc: [] };
}

export const DmSegMobileReply: MessageFns<DmSegMobileReply> = {
  encode(message: DmSegMobileReply, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.elems) {
      DanmakuElem.encode(v!, writer.uint32(10).fork()).join();
    }
    if (message.state !== undefined) {
      writer.uint32(16).int32(message.state);
    }
    if (message.aiFlag !== undefined) {
      DanmakuAIFlag.encode(message.aiFlag, writer.uint32(26).fork()).join();
    }
    for (const v of message.colorfulSrc) {
      DmColorful.encode(v!, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DmSegMobileReply {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDmSegMobileReply();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.elems.push(DanmakuElem.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.state = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.aiFlag = DanmakuAIFlag.decode(reader, reader.uint32());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.colorfulSrc.push(DmColorful.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const DanmakuAIFlag: MessageFns<DanmakuAIFlag> = {
  encode(message: DanmakuAIFlag, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.dmFlags) {
      DanmakuFlag.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DanmakuAIFlag {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDanmakuAIFlag();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.dmFlags.push(DanmakuFlag.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseDanmakuFlag(): DanmakuFlag {
  return { dmid: 0n, flag: 0 };
}

export const DanmakuFlag: MessageFns<DanmakuFlag> = {
  encode(message: DanmakuFlag, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.dmid !== 0n) {
      if (BigInt.asIntN(64, message.dmid) !== message.dmid) {
        throw new globalThis.Error("value provided for field message.dmid of type int64 too large");
      }
      writer.uint32(8).int64(message.dmid);
    }
    if (message.flag !== 0) {
      writer.uint32(16).uint32(message.flag);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DanmakuFlag {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDanmakuFlag();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.dmid = reader.int64() as bigint;
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.flag = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseDmColorful(): DmColorful {
  return { type: 0, src: "" };
}

export const DmColorful: MessageFns<DmColorful> = {
  encode(message: DmColorful, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== 0) {
      writer.uint32(8).int32(message.type);
    }
    if (message.src !== "") {
      writer.uint32(18).string(message.src);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DmColorful {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDmColorful();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.src = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
}